refactor: move queries to respective modules for improved import organization
