+++ src/backend/src/modules/game/queries/deleteGame.js
