+++ src/backend/src/modules/game/queries/findGamesByPlatformId.js
