+++ src/backend/src/modules/platform/dto/createPlatformDto.js
@@ -1,5 +1,4 @@
 /**
  * @typedef {Object} CreatePlatformDto
  * @property {string} name
-
  */
