+++ src/backend/src/modules/platform/queries/createPlatform.js
