+++ src/backend/src/modules/platform/queries/deletePlatform.js
