+++ src/backend/src/modules/platform/queries/findOnePlatform.js
