+++ src/backend/src/modules/platform/queries/findOnePlatformByName.js
