+++ src/backend/src/modules/platform/queries/findPlatforms.js
