+++ src/backend/src/modules/platform/queries/updatePlatform.js
