+++ src/backend/src/modules/user/queries/deleteUser.js
