+++ src/backend/src/modules/user/queries/findOneByEmail.js
