+++ src/backend/src/modules/user/queries/findOneByUsername.js
